ParameterDirectivesTest fixed
